Bug/Cyclic Dependency
- Factored out get_ssm function from utils
